changed the type of input param id of the method existsById for PostRepository from int to long

+++ src/main/java/com/example/blog/dao/PostRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface PostRepository extends JpaRepository<Post, Long> {
 
-    boolean existsById(int id);
+    boolean existsById(long id);
 
     Post findById(long id);
